Moved rratrap input functions to spio.py
